Support binops
This adds back support for code like:

    v && something.macro!()

The issue here was that when parsing the possibly chained
expression, we didn't know what the & meant. It can be either
part of the binary operator in the form of && or of &,
or be part of the expression, by being a unary reference operator:

let v = &something.macro!();

Same holds for - negation/subtraction and * dereference/multiply.