Add repo links, description and README link to Cargo.tomls
postfix-macros-impl doesn't get a README link because
there is no readme for the crate specifically,
and the README in the root was written from the
perspective of the root crate.

It's an annoying pattern that some crates.io
sub-workspace crates link to parent READMEs
that don't have anything to do with the crates
themselves at all.